Backend/requirements.txt
asyncpg==0.30.0
bcrypt==4.2.1
fastapi==0.115.5
pydantic==2.10.2
pydantic_core==2.27.1
SQLAlchemy==2.0.36
uvicorn==0.32.1
passlib==1.7.4
python-jose==3.3.0
numpy
implicit